requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
astor==0.8.1
attrs==25.3.0
audioread==3.0.1
av==15.1.0
babel==2.17.0
bangla==0.0.5
bce-python-sdk==0.9.46
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.40.41
botocore==1.40.41
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
ctranslate2==4.6.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.4
dateparser==1.1.8
decorator==5.2.1
defusedxml==0.7.1
Distance==0.1.3
docopt==0.6.2
easyocr==1.7.2
einops==0.8.1
elevenlabs==2.16.0
encodec==0.1.1
faster-whisper==1.2.0
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
funcy==2.0
future==1.0.0
g2p-en==2.1.0
g2pkk==0.1.2
grpcio==1.75.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
h5py==3.14.0
hangul-romanize==0.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
modelscope==1.30.0
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.1
mss==10.1.0
multidict==6.6.4
murmurhash==1.0.13
networkx==3.5
ninja==1.13.0
nltk==3.9.1
num2words==0.5.14
numba==0.62.1
numpy==2.2.6
ollama==0.6.0
onnx==1.19.0
onnxruntime==1.23.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==25.0
paddlepaddle-gpu==2.6.1
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
preshed==3.0.10
prettytable==3.16.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
py-espeak-ng==0.1.8
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
PyMuPDF==1.26.4
pynndescent==0.5.13
pyparsing==3.2.5
pypdfium2==4.30.0
pyperclip==1.11.0
pypinyin==0.55.0
pyreadline3==3.5.4
PyRect==0.2.0
pysbd==0.3.4
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-doctr==1.0.0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
selenium==4.35.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20250825
SudachiPy==0.6.10
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.9.30
tokenizers==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
trainer==0.0.36
transformers==4.31.0
trio==0.30.0
trio-websocket==0.12.2
TTS==0.22.0
typeguard==4.4.4
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
umap-learn==0.5.9.post2
Unidecode==1.4.0
urllib3==2.5.0
validators==0.35.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
